requirements.txt
certifi
charset-normalizer
idna
pip
python-dotenv
requests
slack_sdk
urllib3